+++ src/radiology-tabs/radiology-tabs.component.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { TabPanels, TabList, Tabs, Tab, TabPanel } from "@carbon/react";
 import { TestsOrdered } from "./test-ordered/tests-ordered.component";
 import { useTranslation } from "react-i18next";
-import WorkList from "../work-list/work-list.component";
+import WorkList from "./work-list/work-list.component";
 import { ReferredTests } from "./referred-test/referred-ordered.component";
 import { Review } from "./review/review-ordered.component";
 import { ApprovedOrders } from "./approved/approved-orders.component";
