+++ src/radiology-tabs/work-list/work-list.resource.ts
