+++ src/radiology-tabs/work-list/work-list.scss
@@ -1,7 +1,7 @@
 @use '@carbon/styles/scss/spacing';
 @use '@carbon/styles/scss/type';
 @import "~@openmrs/esm-styleguide/src/vars";
-@import '../root.scss';
+@import '../../root.scss';
 
 title {
   width: 6.938rem;
