+++ src/root.scss
@@ -1,0 +1,50 @@
+@use '@carbon/styles/scss/spacing';
+@use '@carbon/styles/scss/type';
+@import '~@openmrs/esm-styleguide/src/vars';
+
+.productiveHeading01 {
+    @include type.type-style("heading-01");
+}
+
+.productiveHeading02 {
+    @include type.type-style("heading-02");
+}
+
+.productiveHeading03 {
+    @include type.type-style("heading-03");
+}
+
+.productiveHeading04 {
+    @include type.type-style("heading-04");
+}
+
+.bodyLong01 {
+    @include type.type-style("body-01");
+}
+
+.bodyShort01 {
+    @include type.type-style("body-compact-01");
+}
+
+.bodyShort02 {
+    @include type.type-style("body-compact-02");
+}
+
+.text02 {
+    color: $text-02;
+}
+
+.label01 {
+    @include type.type-style("label-01")
+}
+
+.modal {
+    position: fixed;
+    width: 100vw;
+    height: 100vh;
+    display: grid;
+    background-color: rgba(0, 0, 0, 0.5);
+    z-index: 9000;
+    justify-items: center;
+    align-items: center;
+}
